기능:dualport task 수정 켈리브레이션 방식 개선

+++ Application/App_drivers/app_adc.h
@@ -45,10 +45,10 @@ typedef enum adc_diff_ch_e
 
 
 void adc_init(void);
-int32_t adc_read_single(int channel,uint8_t *err);
+float adc_read_single(int channel,uint8_t *err);
 int32_t adc_read_diff(int channel,uint8_t *err);
 float cvt_adcToVol(int32_t adc,int32_t off,int32_t full,int32_t off_in,int32_t full_in);
-int32_t adc_read_single_avg(int channel,uint8_t *err,uint8_t avg_cnt);
+float adc_read_single_avg(int channel,uint8_t *err,uint8_t avg_cnt);
 int32_t adc_read_diff_avg(int channel,uint8_t *err,uint8_t avg_cnt);
 
 float adc_chToVoltage(int32_t mode,int32_t channel,int32_t adc);
@@ -63,7 +63,7 @@ int32_t get_adc_single_fullset(int channel);
 float adc_read_volate_single(int32_t ch,uint8_t *err);
 float calculate_voltage(adc_config_t *adc_config,uint8_t *err);
 
-int32_t adc_read_single_raw(int channel, uint8_t *err);
+float adc_read_single_raw(int channel, uint8_t *err);
 
 int32_t adc_read_diff_raw(int channel, uint8_t *err);
 #endif
